refactor: modularize player functionality and improve movement and interaction handling
